// aoo.front/src/models/ProjectDetails.ts
export type ProjectDetails = {
    project_id: number;
    main_sme: string;
    backup_sme: string;
    regional_backup: string;
    pem: string;
    functional_ss_poc: string;
    technical_ss_poc: string;
    partner: string;
    t_shirt_size: string;
    start_date: string;
    go_live_date: string;
    sign_off_date: string;
    customer_id: number;
};